refactor: single owner for client-backed event-info lookup helpers
thread_membership and thread_room_scan each carried an identical
_event_info_from_lookup_response normalizer and an identical
conversation-cache thread-id lookup. Move fetch_event_info_for_client
and lookup_thread_id_from_conversation_cache to thread_room_scan (which
already owned the cache-backed variant) and delete the duplicates, so
lookup-response normalization has exactly one owner and
thread_membership becomes pure domain logic with no nio transport
reads. Update the thread_room_scan Tach interface accordingly.
