+++ tach.toml
@@ -2827,7 +2827,9 @@ exclusive = true
 [[interfaces]]
 expose = [
     "RoomScanConversationCache",
+    "fetch_event_info_for_client",
     "fetch_event_info_from_conversation_cache",
+    "lookup_thread_id_from_conversation_cache",
     "room_scan_membership_access_for_client",
 ]
 from = ["mindroom.matrix.thread_room_scan"]
